+++ tic.R
@@ -1,5 +1,9 @@
 do_package_checks(error_on = "warning")
 
+# for parallelLibrary() test
+get_stage("install") %>%
+  add_step(step_install_cran("rpart"))
+
 if (ci_has_env("BUILD_PKGDOWN")) {
   get_stage("install") %>%
     add_step(step_install_github("mlr-org/mlr3pkgdowntemplate"))